chore: move all rule tests and files together
