--- src/rules/createRule.ts
@@ -1,3 +1,0 @@
-import { ESLintUtils } from '@typescript-eslint/utils'
-
-export const createRule = ESLintUtils.RuleCreator(() => '')
