+++ src/rules/tests/setup/tsconfig.json
